LayoutTests/platform/chromium/fast/events/touch/gesture/touch-gesture-scroll-page-propagated-expected.txt

This tests that a gesture scroll is propagated from a div to the page when the div has no remaining scroll offset.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


first gesture
PASS document.body.scrollTop is 0
PASS document.body.scrollLeft is 0
PASS wheelEventsOccurred is 0
second gesture
PASS document.body.scrollTop is 160
PASS document.body.scrollLeft is 0
PASS wheelEventsOccurred is 0
scroll event 0+> [object HTMLDivElement]
PASS scrollEventsOccurred is 1
PASS successfullyParsed is true

TEST COMPLETE
